requirements.txt
tensorflow
opencv-python
numpy
pandas
pyyaml
matplotlib
scikit-learn
Pillow
